requirements.txt
paho-mqtt==1.5.0
requests==2.21.0
pyyaml==5.3
psutil==5.6.7
